+++ app/components/FloatingBar/index.tsx
@@ -17,16 +17,21 @@ interface FloatingBarProps {
   onCircleClick: (event: any) => any;
   onSelectClick: (event: any) => any;
   onPolygonClick: (event: any) => any;
+  onInvisibleClick: (event: any) => any;
   activeButton: SupportedShapes;
 }
 
 const FloatingBar: React.SFC<FloatingBarProps> = (props: FloatingBarProps) => {
-
   const buttons = [{
     icon: icons.select,
     tool: SupportedShapes.selector,
     event: props.onSelectClick,
     helpMessage: 'select (shortcut: escape)',
+  }, {
+    icon: icons.select,
+    tool: SupportedShapes.invisible,
+    event: props.onInvisibleClick,
+    helpMessage: 'Create an empty annotation (shortcut: n)',
   }, {
     icon: icons.anchorRectangle,
     tool: SupportedShapes.rectangle,
