+++ app/containers/Editor/constants.ts
@@ -7,7 +7,6 @@
 enum ActionTypes {
   DEFAULT_ACTION = 'app/Editor/DEFAULT_ACTION',
   CREATE_ANNOTATION = 'app/Editor/CREATE_ANNOTATION',
-  CREATE_INVISIBLE_ANNOTATION = 'app/Editor/CREATE_INVISIBLE_ANNOTATION',
   REMOVE_ANNOTATION = 'app/Editor/REMOVE_ANNOTATION',
   EDIT_ANNOTATION = 'app/Editor/EDIT_ANNOTATION',
   CHANGE_ORDER = 'app/Editor/CHANGE_ORDER',
